Reorganise fastapi app to support multiple webhooks.
I plan to add support for forgejo and it will also just be a simple hook
and we need an API endpoint for it. It makes sense to run just one
lambda and have different paths for gitlab and forgejo.
