+++ .github/workflows/deploy.yml
@@ -62,19 +62,19 @@ jobs:
             ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ env.AWS_DEFAULT_REGION }}.amazonaws.com/sbobot/bot:github-app-latest
             ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ env.AWS_DEFAULT_REGION }}.amazonaws.com/sbobot/bot:github-app-${{ env.DOCKER_IMAGE_VERSION }}
 
-      - name: Build and push gitlab-webhook
+      - name: Build and push webhook-api
         uses: docker/build-push-action@f9f3042f7e2789586610d6e8b85c8f03e5195baf # v7.2.0
         with:
           push: true
           platforms: linux/arm64/v8
-          cache-from: type=gha,scope=gitlabwebhook
-          cache-to: type=gha,mode=max,scope=gitlabwebhook
-          context: gitlab
-          file: gitlab/Dockerfile
+          cache-from: type=gha,scope=webhook-api
+          cache-to: type=gha,mode=max,scope=webhook-api
+          context: webhook-api
+          file: webhook-api/Dockerfile
           provenance: false
           tags: |
-            ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ env.AWS_DEFAULT_REGION }}.amazonaws.com/sbobot/bot:gitlab-webhook-latest
-            ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ env.AWS_DEFAULT_REGION }}.amazonaws.com/sbobot/bot:gitlab-webhook-${{ env.DOCKER_IMAGE_VERSION }}
+            ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ env.AWS_DEFAULT_REGION }}.amazonaws.com/sbobot/bot:webhook-api-latest
+            ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ env.AWS_DEFAULT_REGION }}.amazonaws.com/sbobot/bot:webhook-api-${{ env.DOCKER_IMAGE_VERSION }}
 
       - name: Configure tofu plugin cache
         run: |
@@ -121,7 +121,7 @@ jobs:
             aws lambda delete-function --function-name "$name" --qualifier "$version"
           done
 
-          name="sbobot-gitlab-webhook"
+          name="sbobot-webhook-api"
           aws lambda list-versions-by-function --function-name "$name" | jq -r '.Versions | .[] | [select(.Version != "$LATEST")] | .[].Version' | tac | sed '1,10d' | tac | while read -r version ; do
             printf 'Removing version %s of function %s\n' "$version" "$name"
             aws lambda delete-function --function-name "$name" --qualifier "$version"
@@ -148,7 +148,7 @@ jobs:
         with:
           version: ${{ env.UV_VERSION }}
           enable-cache: true
-          cache-dependency-glob: 'gitlab/uv.lock'
+          cache-dependency-glob: 'webhook-api/uv.lock'
 
       - name: Configure pre-commit repo cache
         uses: actions/cache@27d5ce7f107fe9357f9df03efb73ab90386fccae # v5.0.5
@@ -184,10 +184,10 @@ jobs:
 
       - uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405 # v6.2.0
         with:
-          python-version-file: 'gitlab/.python-version'
+          python-version-file: 'webhook-api/.python-version'
 
       - name: Install python dependencies
-        working-directory: gitlab
+        working-directory: webhook-api
         run: |
           uv sync --all-extras --dev --frozen
