+++ .pre-commit-config.yaml
@@ -48,33 +48,33 @@ repos:
       - id: ruff-lint
         name: ruff-lint
         entry:
-          bash -c 'cd gitlab && uv run ruff check --fix --exit-non-zero-on-fix
-          --force-exclude'
+          bash -c 'cd webhook-api && uv run ruff check --fix
+          --exit-non-zero-on-fix --force-exclude'
         language: python
         require_serial: true
         args: []
         types_or: [python, pyi]
         additional_dependencies: []
-        files: ^gitlab/
+        files: ^webhook-api/
 
       - id: ruff-format
         name: ruff-format
-        entry: bash -c 'cd gitlab && uv run ruff format --force-exclude'
+        entry: bash -c 'cd webhook-api && uv run ruff format --force-exclude'
         language: python
         require_serial: true
         args: []
         types_or: [python, pyi]
         additional_dependencies: []
-        files: ^gitlab/
+        files: ^webhook-api/
 
       - id: pyright
         name: pyright
-        entry: bash -c 'unset VIRTUAL_ENV && cd gitlab && uv run pyright'
+        entry: bash -c 'unset VIRTUAL_ENV && cd webhook-api && uv run pyright'
         language: python
         require_serial: true
         types_or: [python, pyi]
         additional_dependencies: []
-        files: ^gitlab/
+        files: ^webhook-api/
 
       # - id: e2e-tests-eslint
       #   name: eslint
